+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## []
+### Added
+- Input CLI parameter `--config` to specify a configuration file
+
 ## [2.0.0] - 2020-07-29
 ### Changed
 - Improved and expanded configuration file format